roachtest: refactor openmetrics utils and labels
1. Create new `metrics_utils.go` and refactor all openmetrics related
utils into this function
2. Add `Owner()` function in test interface.
3. Remove `subtest` labels since they are making the dashboard labels
confusing
4. Add new flag `openmetrics-labels` in roachtest which
currently adds `branch, arch, suite` labels

Epic: https://cockroachlabs.atlassian.net/browse/CRDB-41852

Release note: None
